refactor: move GetUsers tests to __tests__ folder
